Feat: 슬라이더 범용 컴포넌트로 변경
