+++ app/test/index.tsx
@@ -1,4 +1,4 @@
-import Slider from "@/src/widgets/slide/slider2";
+import Slider from "@/src/widgets/slide";
 import React from "react";
 import { StyleSheet, View } from "react-native";
