+++ src/widgets/index.tsx
@@ -1,9 +1,8 @@
-export * from './checkbox-label';
-export * from './label-input';
-export * from './form';
-export * from './chip-selector';
-export { GemStoreWidget } from './gem-store';
-export * from './slide';
-export * from './step-indicator';
-export * from './ticket';
-
+export * from "./checkbox-label";
+export * from "./label-input";
+export * from "./form";
+export * from "./chip-selector";
+export { GemStoreWidget } from "./gem-store";
+export * from "./slide/index";
+export * from "./step-indicator";
+export * from "./ticket";
